For what it is, this is pretty clean
Got referential attributes working, lifetimes, etc. There are certainly
some rough spots. Using writeln!, dealing with fmt::Error sucks. I
either need a macro, or stop treating the string like a Writer.

Code is clean, has nice spacing, etc. But no indentation. No point if
we are just going to run it through rustfmt.

It's refreshing using the relationship macros to get at data. I don't know
if it's any better than the other way. It's hard to compare because It
was organized oddly. If I think back to ooa_0, I had a dict of objects,
and a dict of relationships. So, this is certainly way better. But as for
the last code generator, I don't know. I like being able to lookup the
relationships I need on the model, like a schematic.

Note that chaining would sure be nice.... But, I do find that I need
data from objects along the way, so there's that.