Done and done!
Got source code tagging outputting. Next up I imagine, will be
formatting, followed by diffing. Diffing was a lot of work, and if
I can leverage any of it, that would be cool. But I don't want to
inherit a hack. And I was rushing things near the end. I just needed
to finish.