rustfmt
Based the implementation off of nut. It's better though. Errors are
routed to vscode. Haven't been able to test one of the big sarzak-
type errors yet. vscode is behind a feature gate. Otherwise it prints
to stderr.

Changed up the builders some. Sending the store to the parent builder.
Super proud of how this is turning out.