This seems to be mostly working
I didn't add the escape hatch.

It seems to work, but I'm really not that pleased with the results. I
think that there is a better way to handle this. I think it's not one
line at a time, but rather should be done one section at a time.
Where a section is a directive block.

That's just a notional thought, as I write this. But I may try it out...