Big Commit
And I've committed on this branch already once or more.

Anyway, new is done! And I'm quite pleased with the results. Especially
that it's as modular as it is, but also because it's using the type system
when generating code. I don't feel like calling foo.id to get a UUID is
a hack. It's type dependent.

The code needs some serious refactoring, and woog needs updating
with new objects. But this has been a big push, and it's finally done!