Giving up for now.
This doesn't even compile. I'm breaking my own rule about that. Well,
as long as I don't push it to the repo, I suppose it's ok.

I may very well pick this up tomorrow. But I really need the option to
be free of it. It has been blocking me, and I shouldn't let something
block me. I should move on. So that's what I'm doing.

I was on the path to implementing something that should work.
Basically an external entity type in the woog domain that contained
all the compiler specific stuff like an lvalue, and a boolean to indicate
if it's been initialized. I'd set this up before each method invocation
and be off to the races.

I was modeling it, and the woog model is a mess. I need to refactor the
💩 out of it. I however just can't face having to go through everything
to make it generate properly, just to maybe change it again down
the road? I guess I could fix nut. No, nut was a means to an end.

I realize now that I can just hand craft the missing items. Maybe not
refactor the model, but put them in todo, and move forward with this.
It is really close, I think...