Finally, some progress!
One to one relationships are done! Found bugs in nut/cuckoo too.
I double checked options, and other shit. Put the functions on the
struct, which TBH, is much nice to deal with.

Also, we are always returning a vec from the rel nav funcs! That's
super exciting, because it means these things should be composable.

🍀#13