Added description coloring
It's rather clever. Used the ConfigValue stuff from nut. It really was a
good design. I learned a new trick thanks to copilot. From<Tuple>.
Very cool. Used it to create a configuration type to pass around
the compiler. It takes what came from the options and adds what
is in the description.

I may want to consider how to combine the two sources, rather than
defaulting to what's in the description. Maybe not.

Work towards issue #27.