Imports
Changed imports to use a top-level option to indicate the module
being imported. That's used to add the foreign ObjectStore to the
external entities. Later the object's description is parsed for details,
that we'll use later to inspect the imported object during code
generation, e.g., to determine if it's a supertype.