Well, so there it is...
I did two distinct things. On one hand I changed 1-? to 1-* because
even though I understand ? to mean one of {1,M}, perhaps not everyone
does. Hell, even when I see it I see a question mark. Splat seems
better.

The rest is my initial attempt at getting the doc tests generated.
I gave up. I could probably make it work, but as it is stands there
are a hodgepodge of structs that are duplicated between todo
and woog. The majority of my code uses todo. Moving all that
over to woog is on the plan, after v2.

Bottom line, I'd rather do it correctly, than hack it in now. At least the
code that I had worked on is now in the mainline.