Well, that did it.
I can't believe this is actually done! Seriously. I think I only started
yesterday. I updated the model, and did all the refactoring. That's
not bad TBH. It helps that a huge chunk of the code is generated.
It's the regularity of generated code that really shines here. I should
write about that.

And I really love it when the compiler finds bugs in the model. It
usually does this with a panic, so that's not awesome. Just created
#43 for this.

I've rambled long enough. This closes #41.