Last changes to compile new woog domain
So this was mostly about testing how an object was generated, so
that we can in turn generate the correct thing. This involved testing
against imported objects as well. This was the cause of quite a lot of
consternation for a time.

There are still some warnings extant, but I still need to switch over
to the new woog, so they will be taken care of.