Checkpoint
So, I'm just trying to compare timestamps between what's loaded
from the objectstore and when the file was emitted. The latter comes
from the woog domain. I'm trying to update said domain with a time
stamp, and suddenly I need to support external entities! Holy fuck.

So, I'm mid-process adding that, and guess what? I need to add a
completely new file generator for this. So, on one hand, duh, but on
the other, what I'm trying to accomplish is so utterly simple.

The joys of code generation?