Need to change tack
I'm at the point that I need to output the new method's guts. For the
external entity type. And I don't want to. It still uses the old cruft,
and I am not going to go out of my way to make it work, only to have
to change it tomorrow. Or the next day.

So I'm stopping here and I'm going to take care of this problem. I'll
be back.