Sorting and comments
I'm adding comments. Thinking out loud, as it were. Whilst doing so
I noticed that I wasn't preserving the ordering of the fields in the
structure I was working on. I'm not convinced that they are properly
sorted when they are coming in, but I'm most certainly going over
them in order now.