Putting a pin in it
This is going to be a complete rewrite.

I'd like to see the structural code relegated to a template. The functions
on the other hand should be generated by walking an AST, probably
starting at Block. I think it'll take a set of templates, and an AST
walker for each target language, assuming woog isn't too rust-centric.

I'm not at all sure how things will be processed. I anticipate targets
like we already have. Woog population is generic. So I guess we'll
start there, and populate woog. Now, some of that depends on imports,
which we don't have until we get to the target. At least that's how it
is currently.

I'd really like to see woog get populated prior to target selection. Well,
maybe that's not going to happen. I'm thinking about a particular target
having different function implementations. I don't think it's a problem.
If the function is in woog, then it can be generated, and the target can
decide whether or not to do so. If you want something on the target
that isn't in woog, then it can't be generated anyway, and you may as
well hardcode the thing somehow.

So that's v3 in a nutshell. The branch is called recurse, which was
named before I decided to rewrite, but it works well enough. Sarzak
also has a recurse branch, which contains a branch of woog.

Wish me luck. May the force be with me. Live long and prosper. 🖖