Change uuid to v4
I am having a terrible time tracing a bug in the feature/recurse branch.
I really think it's an aliasing issue, so I'm going to force something to
happen.

I've been meaning to do this. Did do it in the other branch. So it's
not just to debug the problem. Adding a "`salt`" attribute to objects
just to avoid aliasing is dumb.