Just a few things
I changed the hashmap implementation to be `rustc_hash`. I also got
"no deadlocks" working with dwarf. Some random import stuff. I think
that it has to do with not adding all the "uber decorations" around
a parameter that is an imported object. Basically that's for sarzak,
since it's not uber. This is clearly a hack.