Lot's 'o changes
Storing objects as RC in the Proxy type, so now changes in dwarf are
replicated in the store. Removed indirect calls to objectstore, since
we have the store. Added a store to each proxy so that it can do stuff
with it directly. That of course required wrapping it up in an RC<RefCell<>>.